+++ example/lib/main.dart
@@ -8,11 +8,10 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'dart:async';
-
 import 'package:flutter/services.dart';
-import 'package:twocheckout_flutter/model/PaymentFormResult.dart';
-import 'package:twocheckout_flutter/model/PaymentMethodType.dart';
-import 'package:twocheckout_flutter/model/TokenResult.dart';
+import 'package:twocheckout_flutter/model/payment_form_result.dart';
+import 'package:twocheckout_flutter/model/payment_method_type.dart';
+import 'package:twocheckout_flutter/model/token_result.dart';
 import 'package:twocheckout_flutter/twocheckout_flutter.dart';
 import 'package:twocheckout_flutter_example/payment_flow_done.dart';
 
@@ -39,7 +38,7 @@ class _MyAppState extends State<MyApp> implements TwoCheckoutFlutterEvents {
         TwoCheckoutFlutterEventsImpl(twoCheckoutFlutterEvents: this);
 
     /// Set 2Checkout credentials
-    _twoCheckoutFlutterPlugin.setTwoCheckoutCredentials(secretKey: "Your_Key", merchantCode: "Your_Code");
+    _twoCheckoutFlutterPlugin.setTwoCheckoutCredentials(secretKey: "", merchantCode: "");
   }
 
   void showPaymentMethods() {
