+++ pubspec.yaml
@@ -7,7 +7,7 @@ issue_tracker: https://github.com/DevCrew-io/2checkout-flutter/issues
 documentation: https://github.com/DevCrew-io/2checkout-flutter/blob/main/README.md
 
 environment:
-  sdk: '>=3.1.0-231.0.dev <4.0.0'
+  sdk: '>=3.0.0<4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
